.cursor/rules/utils/requirements.txt
# Core dependencies
pyyaml>=6.0
pathlib>=1.0.1

# Code analysis tools
radon>=5.1.0
ast-comments>=1.0.0

# Data processing and visualization
matplotlib>=3.5.0
plotly>=5.10.0

# Documentation generation
markdown>=3.4.0
jinja2>=3.0.0

# Testing and development
pytest>=7.0.0
black>=22.0.0

# Web interface (if needed)
flask>=2.0.0

# Vector search (if needed)
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

colorama>=0.4.6
argparse>=1.4.0
